Create RealTripsStore and provide KrailDB using Dagger

+++ feature/sydney-trains/database/api/src/main/kotlin/xyz/ksharma/krail/database/sydney/trains/database/api/TripsStore.kt
@@ -30,10 +30,10 @@ interface TripsStore {
         tripId: String,
         tripHeadsign: String?,
         tripShortName: String?,
-        directionId: Int?,
+        directionId: Long?,
         blockId: String?,
         shapeId: String?,
-        wheelchairAccessible: Int?
+        wheelchairAccessible: Long?
     )
 
     /**
